requirements.txt
fastapi
uvicorn[standard]
pydantic
cfbd
python-dotenv